requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.5.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
accelerate=1.3.0=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiohttp-cors=0.7.0=pypi_0
aiosignal=1.3.2=pypi_0
airportsdata=20241001=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
astor=0.8.1=pypi_0
asttokens=3.0.0=pypi_0
async-lru=2.0.4=pypi_0
attrs=25.1.0=pypi_0
babel=2.17.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.13.3=pypi_0
bitsandbytes=0.45.1=pypi_0
blake3=1.0.4=pypi_0
bleach=6.2.0=pypi_0
brotli=1.1.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.7.14=hbd8a1cb_0
cachetools=5.5.1=pypi_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
cloudpickle=3.1.1=pypi_0
colorful=0.5.6=pypi_0
comm=0.2.2=pypi_0
compressed-tensors=0.9.0=pypi_0
contourpy=1.3.2=pypi_0
cos-python-sdk-v5=1.9.33=pypi_0
crcmod=1.7=pypi_0
cut-cross-entropy=25.1.1=pypi_0
cycler=0.12.1=pypi_0
datasets=3.1.0=pypi_0
debugpy=1.8.12=pypi_0
decorator=5.1.1=pypi_0
deepspeed=0.15.4=pypi_0
defusedxml=0.7.1=pypi_0
depyf=0.18.0=pypi_0
diffusers=0.34.0=pypi_0
dill=0.3.8=pypi_0
diskcache=5.6.3=pypi_0
distlib=0.3.9=pypi_0
distro=1.9.0=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
docstring-parser=0.16=pypi_0
einops=0.8.0=pypi_0
executing=2.2.0=pypi_0
expat=2.7.1=h6a678d5_0
fastapi=0.115.8=pypi_0
fastjsonschema=2.21.1=pypi_0
filelock=3.17.0=pypi_0
flash-attn=2.7.4.post1=pypi_0
fonttools=4.58.5=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.9.0=pypi_0
gguf=0.10.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
gmpy2=2.2.1=pypi_0
google-api-core=2.24.1=pypi_0
google-auth=2.38.0=pypi_0
googleapis-common-protos=1.66.0=pypi_0
grpcio=1.70.0=pypi_0
h11=0.14.0=pypi_0
hf-transfer=0.1.9=pypi_0
hjson=3.1.0=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.28.1=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.6.1=pypi_0
iniconfig=2.0.0=pypi_0
intel-cmplr-lib-ur=2025.2.0=pypi_0
intel-openmp=2025.2.0=pypi_0
interegular=0.3.3=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.12.3=pypi_0
ipywidgets=8.1.5=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.4=pypi_0
jiter=0.8.2=pypi_0
joblib=1.5.1=pypi_0
json5=0.10.0=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter=1.1.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.12.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.15.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.3.5=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.8=pypi_0
lark=1.2.2=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc=15.1.0=h767d61c_3
libgcc-ng=15.1.0=h69a702a_3
libgomp=15.1.0=h767d61c_3
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
lm-format-enforcer=0.10.9=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.3=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mistral-common=1.5.2=pypi_0
mistune=3.1.1=pypi_0
mkl=2025.2.0=pypi_0
mkl-fft=1.3.11=pypi_0
mkl-random=1.2.8=pypi_0
mkl-service=2.4.2=pypi_0
modelscope=1.22.3=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
msgspec=0.19.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.5=h2d0b736_3
nest-asyncio=1.6.0=pypi_0
networkx=3.3=pypi_0
ninja=1.11.1.3=pypi_0
nltk=3.9.1=pypi_0
notebook=7.3.2=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-ml-py=12.570.86=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
nvitop=1.5.1=pypi_0
openai=1.61.0=pypi_0
opencensus=0.11.4=pypi_0
opencensus-context=0.1.3=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openssl=3.5.1=h7b32b05_0
outlines=0.1.11=pypi_0
outlines-core=0.1.26=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
partial-json-parser=0.2.1.1.post5=pypi_0
peewee=3.17.8=pypi_0
peft=0.14.0=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.4.0=pypi_0
pip=25.1=pyhc872135_2
pipreqs=0.5.0=pypi_0
platformdirs=4.3.6=pypi_0
pluggy=1.5.0=pypi_0
prometheus-client=0.21.1=pypi_0
prometheus-fastapi-instrumentator=7.0.2=pypi_0
prompt-toolkit=3.0.50=pypi_0
propcache=0.2.1=pypi_0
proto-plus=1.26.0=pypi_0
protobuf=3.20.3=pypi_0
psutil=6.1.1=pypi_0
pthread-stubs=0.3=h0ce48e5_1
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
py-cpuinfo=9.0.0=pypi_0
py-spy=0.4.0=pypi_0
pyarrow=19.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pybind11=2.13.6=pypi_0
pycountry=24.6.1=pypi_0
pycparser=2.22=pypi_0
pycryptodome=3.21.0=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.19.1=pypi_0
pyjwt=2.8.0=pypi_0
pynvml=12.0.0=pypi_0
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=pypi_0
pytest=8.3.4=pypi_0
python=3.11.13=h1a3bd86_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-json-logger=3.2.1=pypi_0
pytz=2025.1=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.1=pypi_0
rank-bm25=0.2.2=pypi_0
ray=2.41.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.22.3=pypi_0
rsa=4.9=pypi_0
safetensors=0.5.2=pypi_0
scikit-learn=1.7.0=pypi_0
scipy=1.16.0=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.3=pypi_0
sentence-transformers=5.0.0=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.20.0=pypi_0
setproctitle=1.3.4=pypi_0
setuptools=75.8.0=pypi_0
shtab=1.7.2=pypi_0
six=1.17.0=pypi_0
smart-open=7.1.0=pypi_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
socksio=1.0.0=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
starlette=0.45.3=pypi_0
swanboard=0.1.7b1=pypi_0
swankit=0.1.2b6=pypi_0
swanlab=0.4.6=pypi_0
sympy=1.13.1=pypi_0
tbb=2022.2.0=pypi_0
tcmlib=1.4.0=pypi_0
tensorboard=2.18.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.6.0=pypi_0
tiktoken=0.7.0=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.14=h993c535_1
tokenizers=0.21.0=pypi_0
torch=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.48.1=pypi_0
triton=3.1.0=pypi_0
trl=0.15.2=pypi_0
typeguard=4.4.4=pypi_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing-extensions=4.14.0=pypi_0
tyro=0.9.26=pypi_0
tzdata=2025.1=pypi_0
ujson=5.10.0=pypi_0
umf=0.11.0=pypi_0
unsloth=2025.3.8=pypi_0
unsloth-zoo=2025.4.2=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
uvloop=0.21.0=pypi_0
virtualenv=20.29.1=pypi_0
vllm=0.7.0=pypi_0
wandb=0.19.5=pypi_0
watchfiles=1.0.4=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
websockets=14.2=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py311h06a4308_0
widgetsnbextension=4.0.13=pypi_0
wrapt=1.17.2=pypi_0
xformers=0.0.28.post3=pypi_0
xgrammar=0.1.11=pypi_0
xmltodict=0.14.2=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xxhash=3.5.0=pypi_0
xz=5.6.4=h5eee18b_1
yarg=0.1.9=pypi_0
yarl=1.18.3=pypi_0
zhipuai=2.1.5.20250701=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
